Wrote some test. Ran the tests. Fixed the problems.
Wrote a procedural test.  It adds some events and then triggers them.  Passing is all 12 logging true.

The tests revealed that the trigger function had an issue.  It would only call one set of callbacks if there were events with the same name but different namespaces.  So that is working correctly now.